AEA-277 Start to remove sandbox
