+++ .gitignore
@@ -4,10 +4,9 @@
 **/public/
 **/coverage/
 **/node_modules/
-sandbox/mocks
-sandbox/tests/resources
 coordinator/tests/resources/parent-prescription-1/fhir-message.json
-tests/e2e/electronic-prescription-service-api-sandbox.json
+coordinator/tests/resources/parent-prescription-2/fhir-message.json
+tests/e2e/electronic-prescription-coordinator-postman-tests.json
 
 .#*
