+++ Makefile
@@ -4,9 +4,9 @@ SHELL=/bin/bash -euo pipefail
 
 install: install-node install-python install-hooks
 
-build: build-models build-specification build-sandbox build-coordinator build-proxies
+build: build-models build-specification build-coordinator build-proxies
 
-test: validate-models lint check-licenses test-coordinator #test-sandbox -> TODO: Consolidate coordinator to sandbox, move integration tests out of this target OR remove dep on background process
+test: validate-models lint check-licenses test-coordinator
 
 release:
 	mkdir -p dist
@@ -22,15 +22,13 @@ clean:
 	rm -rf sandbox/tests/resources
 	rm -rf coordinator/dist
 	rm -f coordinator/tests/resources/parent-prescription-1/fhir-message.json
+	rm -f coordinator/tests/resources/parent-prescription-2/fhir-message.json
 
 ## Run
 
 run-specification:
 	scripts/set_spec_server_dev.sh
 	npm run serve
-
-run-sandbox:
-	cd sandbox && npm run start
 
 run-coordinator:
 	cp coordinator/package.json coordinator/dist/
@@ -60,6 +58,7 @@ build-models:
 	$(foreach file, $(wildcard models/requests/*.yaml), cp $(file) models/dist/requests;)
 	$(foreach file, $(wildcard models/requests/*.json), cp $(file) models/dist/requests;)
 	$(foreach file, $(wildcard models/requests/*.yaml), poetry run python scripts/yaml2json.py $(file) models/dist/requests;)
+	$(foreach file, $(wildcard models/responses/*.xml), cp $(file) models/dist/responses;)
 	$(foreach file, $(wildcard models/responses/*.yaml), cp $(file) models/dist/responses;)
 	$(foreach file, $(wildcard models/responses/*.json), cp $(file) models/dist/responses;)
 	$(foreach file, $(wildcard models/responses/*.yaml), poetry run python scripts/yaml2json.py $(file) models/dist/responses;)
@@ -86,33 +85,28 @@ build-specification:
 
 build-coordinator:
 	cp models/dist/requests/PrepareSuccessRequest.json coordinator/tests/resources/parent-prescription-1/fhir-message.json
-	cd coordinator \
-	&& npm run build
-
-build-sandbox:
-	mkdir -p sandbox/mocks
-	cp -r models/dist/responses/*.json sandbox/mocks
-	mkdir -p sandbox/tests/resources
-	cp models/dist/requests/PrepareSuccessRequest.json sandbox/tests/resources/valid-bundle.json
-	cp models/dist/requests/SendSuccessRequest.json sandbox/tests/resources/valid-bundle-with-signature.json
-	cp -r models/dist/responses/*.json sandbox/mocks/
-	poetry run scripts/update_sandbox_tests.py
+	cp models/dist/requests/PrepareSuccessNominatedPharmacyRequest.json coordinator/tests/resources/parent-prescription-2/fhir-message.json
+	npm run --prefix=coordinator/ build
+	poetry run scripts/update_coordinator_tests.py
 
 build-proxies:
-	scripts/build_proxies.sh
+	mkdir -p dist/proxies/sandbox
+	mkdir -p dist/proxies/live
+	cp -Rv proxies/sandbox/apiproxy dist/proxies/sandbox
+	cp -Rv proxies/live/apiproxy dist/proxies/live
 
 # Test
-
-test-sandbox:
-	cd sandbox \
-	&& export ENVIRONMENT=$(or $(ENVIRONMENT),local) \
-	&& export API_TEST_ENV_FILE_PATH=$(or $(API_TEST_ENV_FILE_PATH),../tests/e2e/environments/local.postman_environment.json) \
-	&& export API_TEST_URL=$(or $(API_TEST_URL),localhost:9000) \
-	&& npm run test
 
 test-coordinator:
 	cd coordinator \
 	&& npm run test
+
+# Integration Test
+
+test-integration-coordinator:
+	cd coordinator \
+	&& export API_TEST_ENV_FILE_PATH=$(or $(API_TEST_ENV_FILE_PATH),../tests/e2e/environments/local.postman_environment.json) \
+	&& npm run integration-test
 
 ## Quality Checks
 
@@ -122,13 +116,11 @@ validate-models:
 
 lint:
 	cd specification && npm run lint
-	cd sandbox && npm run lint
 	cd coordinator && npm run lint
 	poetry run flake8 scripts/*.py --config .flake8
 	shellcheck scripts/*.sh
 
 check-licenses:
 	cd specification && npm run check-licenses
-	cd sandbox && npm run check-licenses
 	cd coordinator && npm run check-licenses
 	scripts/check_python_licenses.sh
