+++ azure-pipelines.yml
@@ -97,8 +97,8 @@ jobs:
       - bash: "make build"
         displayName: Build
 
-      - bash: "make run-sandbox &"
-        displayName: Start Sandbox Server
+      - bash: "make run-coordinator &"
+        displayName: Start Coordinator
 
       - bash: "make test"
         displayName: Run Quality Checks
@@ -106,8 +106,8 @@ jobs:
       - bash: 'echo "##vso[task.setvariable variable=SPEC_VERSION]$(poetry run python scripts/calculate_version.py)"'
         displayName: Set SPEC_VERSION
 
-      - bash: "make test-sandbox"
-        displayName: Run Sandbox Integration Tests
+      - bash: "make test-integration-coordinator"
+        displayName: Run Coordinator Integration Tests
 
       - task: PublishTestResults@2
         inputs:
