+++ scripts/update_coordinator_tests.py
@@ -1,0 +1,141 @@
+#!/usr/bin/env python3
+"""
+update_coordinator_tests.py
+
+Updates expected e2e integration test json responses and their corresponding requests for the
+coordinator, converting dates correctly.
+
+Usage:
+  update_coordinator_tests.py
+"""
+
+import json
+import datetime
+
+
+def date_converter(obj):
+    """Date and datetime converter to correctly render dates in json"""
+    if isinstance(obj, datetime.datetime):
+        return obj.replace(tzinfo=datetime.timezone.utc).isoformat()
+    if isinstance(obj, datetime.date):
+        return obj.isoformat()
+    return obj
+
+
+def main():
+    """Main entrypoint"""
+    postman_file_path = "./tests/e2e/electronic-prescription-coordinator-postman-tests.json.template" #todo: parametise
+    with open(postman_file_path) as f:
+        data = json.load(f)
+
+    #todo: extract function, parametise
+    convert_success_request_file_path = "./models/dist/requests/ConvertSuccessRequest.json"
+    with open(convert_success_request_file_path) as f:
+        convert_success_request = json.load(f)
+
+    convert_success_response_file_path = "./models/dist/responses/ConvertSuccessResponse.xml"
+    with open(convert_success_response_file_path) as f:
+        convert_success_response = f.read().replace("\n", "\\n")
+
+    prepare_success_request_file_path = "./models/dist/requests/PrepareSuccessRequest.json"
+    with open(prepare_success_request_file_path) as f:
+        prepare_success_request = json.load(f)
+
+    prepare_success_response_file_path = "./models/dist/responses/PrepareSuccessResponse.json"
+    with open(prepare_success_response_file_path) as f:
+        prepare_success_response = json.load(f)
+
+    send_success_request_file_path = "./models/dist/requests/SendSuccessRequest.json"
+    with open(send_success_request_file_path) as f:
+        send_success_request = json.load(f)
+
+    # send_success_response_file_path = "./models/dist/responses/SendSuccessResponse.json"
+    # with open(send_success_response_file_path) as f:
+    #     send_success_response = json.load(f)
+
+    for item in data['item']:
+        for event in item['event']:
+            #todo: extract function, parametise
+            if (event['script']['id'] == "582bca6a-3e80-4609-be0c-1fc7a05d7d34"):
+                item['request']['body']['raw'] = json.dumps(
+                    convert_success_request,
+                    default=date_converter,
+                    separators=(',', ':'))
+
+                # response = ["const responseString = '"]
+                # for i, line in enumerate(convert_success_response):
+                #     if (i == 0):
+                #         response[i] = response[i] + line + ","
+                #     elif (i == len(convert_success_response) - 1):
+                #         response[i-1] = line + "'"
+                #     else:
+                #         response.append(line)
+
+                # response.extend([
+                #     "const responseString = '" + convert_success_response + "'",
+                #     "pm.test(\"Status code is 200\", function () {",
+                #     "    pm.response.to.have.status(200);",
+                #     "});",
+                #     "pm.test(\"Body is correct\", function () {",
+                #     "    pm.response.to.have.body(responseString);",
+                #     "});"
+                # ])
+
+                # response = ""
+                # for line in convert_success_response:
+                #     response = response + line
+                # response = response.replace("\n", "\\n")
+
+                event['script']['exec'] = [
+                    "const responseString = '" + convert_success_response + "'",
+                    "pm.test(\"Status code is 200\", function () {",
+                    "    pm.response.to.have.status(200);",
+                    "});",
+                    "pm.test(\"Body is correct\", function () {",
+                    "    pm.response.to.have.body(responseString);",
+                    "});"
+                ]
+            elif (event['script']['id'] == "630e7726-f2e1-4bf9-a90f-08350d24e70d"):
+                item['request']['body']['raw'] = json.dumps(
+                    prepare_success_request,
+                    default=date_converter,
+                    separators=(',', ':'))
+
+                event['script']['exec'] = [
+                    "const responseString = '" +
+                    json.dumps(prepare_success_response, default=date_converter, separators=(',', ':'))
+                    + "'",
+                    "pm.test(\"Status code is 200\", function () {",
+                    "    pm.response.to.have.status(200);",
+                    "});",
+                    "pm.test(\"Body is correct\", function () {",
+                    "    pm.response.to.have.body(responseString);",
+                    "});"
+                ]
+            elif (event['script']['id'] == "400fb7e1-0145-41c8-9523-282c047ee1db"):
+                item['request']['body']['raw'] = json.dumps(
+                    send_success_request,
+                    default=date_converter,
+                    separators=(',', ':'))
+
+                event['script']['exec'] = [
+                    "const responseString = 'Message Sent'",
+                    "pm.test(\"Status code is 200\", function () {",
+                    "    pm.response.to.have.status(200);",
+                    "});",
+                    "pm.test(\"Body is correct\", function () {",
+                    "    pm.response.to.have.body(responseString);",
+                    "});"
+                ]
+
+    with open(
+            "./tests/e2e/electronic-prescription-coordinator-postman-tests.json",
+            "w"
+    ) as out_file:
+        out_file.write(
+            json.dumps(data, default=date_converter, indent=2)
+        )
+
+
+if __name__ == "__main__":
+    main()
