--- tests/e2e/environments/internal_dev.postman_environment.json
@@ -1,14 +1,0 @@
-{
-	"id": "52e27ac1-a717-4a09-8826-c022c8cd86ca",
-	"name": "Internal Dev",
-	"values": [
-		{
-			"key": "base_url",
-			"value": "https://internal-dev.api.service.nhs.uk/electronic-prescriptions",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2020-06-24T15:01:58.326Z",
-	"_postman_exported_using": "Postman/7.27.0"
-}
