--- tests/e2e/environments/postman_environment.json.template
@@ -1,14 +1,0 @@
-{
-	"id": "7eaced88-0e9f-420e-8739-4a6aea7a3688",
-	"name": "{{ENVIRONMENT}}",
-	"values": [
-		{
-			"key": "base_url",
-			"value": "{{API_TEST_URL}}",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2020-04-03T10:56:56.598Z",
-	"_postman_exported_using": "Postman/7.21.1"
-}
